+++ src/Reflex/Spider/Orphans.hs
@@ -2,15 +2,17 @@
 Module: Reflex.Spider.Orphans
 Description: Orphan instances for SpiderTimeline and SpiderHost
 -}
+{-# LANGUAGE CPP #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 module Reflex.Spider.Orphans where
 
+#if MIN_VERSION_reflex(0,6,3)
+#else
 import Reflex
 import Reflex.Spider.Internal
 
--- TODO move this to reflex
 instance NotReady (SpiderTimeline x) (SpiderHost x) where
   notReadyUntil _ = pure ()
   notReady = pure ()
@@ -18,3 +20,4 @@ instance NotReady (SpiderTimeline x) (SpiderHost x) where
 instance HasSpiderTimeline x => NotReady (SpiderTimeline x) (PerformEventT (SpiderTimeline x) (SpiderHost x)) where
   notReadyUntil _ = pure ()
   notReady = pure ()
+#endif